Use byte arrays in MeshPrimitiveDecoderExtensions.cs
We are sadly never gonna have a use for Span<T> outside of specialized cases due to the fact that they are ref structs and are highly restrictive and the fact that defaultecs (the ECS system used in the project that I'm making this repository for) also does not support ref structs for their components due to the fact that ref structs cannot be used as type argument.

Span<T>.ToArray() creates a copy of the array, therefore, using them in this context will result in extra unneccercary copies being made.